Split sick note validator
