+++ src/test/java/org/synyx/urlaubsverwaltung/sicknote/sicknote/SickNoteValidatorTest.java
@@ -9,8 +9,6 @@ import org.springframework.validation.BeanPropertyBindingResult;
 import org.springframework.validation.Errors;
 import org.synyx.urlaubsverwaltung.overlap.OverlapService;
 import org.synyx.urlaubsverwaltung.person.Person;
-import org.synyx.urlaubsverwaltung.sicknote.comment.SickNoteCommentEntity;
-import org.synyx.urlaubsverwaltung.sicknote.comment.SickNoteCommentForm;
 import org.synyx.urlaubsverwaltung.workingtime.WorkingTimeService;
 
 import java.time.Clock;
@@ -166,38 +164,6 @@ class SickNoteValidatorTest {
         final Errors errors = new BeanPropertyBindingResult(sickNote, "sickNote");
         sut.validate(sickNote, errors);
         assertThat(errors.getFieldErrors("endDate").get(0).getCode()).isEqualTo("error.entry.invalidPeriod");
-    }
-
-    @Test
-    void ensureCommentMayNotBeNull() {
-        final SickNoteCommentForm sickNoteCommentForm = new SickNoteCommentForm();
-
-        final Errors errors = new BeanPropertyBindingResult(sickNoteCommentForm, "sickNote");
-        sut.validateComment(sickNoteCommentForm, errors);
-        assertThat(errors.getFieldErrors("text").get(0).getCode()).isEqualTo("error.entry.mandatory");
-    }
-
-    @Test
-    void ensureTooLongCommentIsNotValid() {
-
-        final SickNoteCommentForm sickNoteCommentForm = new SickNoteCommentForm();
-        sickNoteCommentForm.setText("Lorem ipsum dolor sit amet, consetetur sadipscing elitr, "
-            + "sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, "
-            + "sed diam voluptua. At vero eos et accusam et justo duo dolores bla bla");
-
-        final Errors errors = new BeanPropertyBindingResult(sickNoteCommentForm, "sickNote");
-        sut.validateComment(sickNoteCommentForm, errors);
-        assertThat(errors.getFieldErrors("text").get(0).getCode()).isEqualTo("error.entry.tooManyChars");
-    }
-
-    @Test
-    void ensureValidCommentHasNoErrors() {
-        final SickNoteCommentForm sickNoteCommentForm = new SickNoteCommentForm();
-        sickNoteCommentForm.setText("I am a fluffy little comment");
-
-        final Errors errors = new BeanPropertyBindingResult(sickNoteCommentForm, "sickNote");
-        sut.validateComment(sickNoteCommentForm, errors);
-        assertThat(errors.getErrorCount()).isZero();
     }
 
     @Test
